+++ packages/core/src/libs/http/routes/api/v1/media.routes.ts
@@ -15,6 +15,7 @@ import getSingle from "../../../controllers/media/get-single.js";
 import getUploadPartUrls from "../../../controllers/media/get-upload-part-urls.js";
 import getUploadSession from "../../../controllers/media/get-upload-session.js";
 import moveFolder from "../../../controllers/media/move-folder.js";
+import requestDownload from "../../../controllers/media/request-download.js";
 import restoreMultiple from "../../../controllers/media/restore-multiple.js";
 import updateSingle from "../../../controllers/media/update-single.js";
 
@@ -45,6 +46,7 @@ const mediaRoutes = new Hono<LucidHonoGeneric>()
 	.post("/upload-session/:sessionId/complete", ...completeUploadSession)
 	.post("/", ...createSingle)
 	.post("/restore", ...restoreMultiple)
+	.post("/:id/download", ...requestDownload)
 	.post("/:id/share-links", ...createMediaShareLink)
 	.patch("/folders/:id", ...updateSingleFolder)
 	.patch("/:id/move", ...moveFolder)