+++ packages/core/src/services/media/index.ts
@@ -15,6 +15,7 @@ export { default as getStorageUsage } from "./get-storage-usage.js";
 export { default as getUploadPartUrls } from "./get-upload-part-urls.js";
 export { default as getUploadSession } from "./get-upload-session.js";
 export { default as moveFolder } from "./move-folder.js";
+export { default as requestDownload } from "./request-download.js";
 export { default as resolveUrl } from "./resolve-url.js";
 export { default as restoreMultiple } from "./restore-multiple.js";
 export * as strategies from "./strategies/index.js";